refactor: extract connection string validation methods
